remove comment regarding email protections
this paragraph is inaccurate, we do not require a verified email address to perform a password reset, see https://github.com/pypa/warehouse/blob/bd81ae8bae40fcd2736e47b65912db297a993339/warehouse/accounts/models.py#L165-L174 which is the only gate on password reset requests.

the email verification concerns noted in this paragraph are more to do with deliverability and reputation than security and would not have changed the course of this incident.